Create a lock screen widget (aka Today Extension) (#281)
* Initial pass at creating a lock screen widget (aka. "Today Extension")

Changes to Loop:
  1. Created a TodayExtensionDataManager which keeps an eye on DeviceDataManager
     and stores data in TodayExtensionContext
  2. Added Loop to a shared app group

Changes Loop TodayExtension:
  1. New target
  2. Reuses the HUD view from StatusTableViewController
  3. Deserializes TodayExtensionContext data from share app group defaults
  4. Displays basic data (currently only the current BG) in the lock view

* Use MAIN_APP_BUNDLE_IDENTIFIER instead of hardcoding com.loudnate

* Rename the widget to just "Loop" and remove the iOS 10.1 install requirement

* Add basal, reservoir and battery data to the today extension.

In addition, inject some debug data if we're running on a simulator so
that we can easily test the experience.

* Improve the way that TodayExtensionContext handles optional sections

* Clean up the UI

- wrap the StackView in a UIView
- set the dimensions properly
- fix all constraint issues

* Add support for net basal rates in the lock screen widget

Along the way, refactor the lastTempBasal visual calculations out
of StatusViewController and into LoopManager so that it can be shared
with TodayExtensionDataManager.

* Pass sensor info along to the extension.

This makes the alert icon do the right thing.
Do some refactors along the way to clean things up.

* Plumb the "eventual glucose" value through to a subtitle in the widget.

* Internationalize the eventual glucose string properly

* Clean up TodayExtensionContext innards

1. Move the store/retrieve semantics into a NSUserDefaults extension
2. Change TodayExtensionContext to implement RawRepresenable and generally
   follow the semantics in WatchContext
3. Fixed all cases where we were storing non-conformant data, which
   was resulting in serialization errors

related: fixed a minor bug in TodayViewController where we were not
calling completionHandler() reliably.

* Rename the widget and all references to 'Loop Status Extension'

* Update version number to 1.1.2 to match container.

* Rename scheme to match new extension name

* Remove boilerplate comments

* Remove vestigial lastContext code

* Change eventual glucose to be a double value instead of a string, and
plumb the preferred unit through the context as well.

* Squish a few small bugs in context data transfer and with preferred units.

* Correctly deserialize GlucoseTrend.

* Fix busted merge.

* Remove unnecessary imports and frameworks to minimize exposure to
frameworks that do not support application extensions (see #292)

* Change GlucoseHUDView.set() to stop taking GlucoseValue

This is in preparation for a future change where StatusExtensionContext
will stop having knowledge of GlucoseValue so that it can stop pulling
in the LoopKit framework, which is not extension-API-safe.

* Elminate all framework dependencies from Loop Status Extension

LoopKit and InsulinKit frameworks are not extension API safe. Switch
to using common data structures that don't reference those frameworks.

* Resolve blown merge.

* Add HKUnit extension to get preferredMinimumFractionDigits

That extension to HKUnit comes from LoopKit, which is no longer
included with the Loop Status Extension. Not sure how this compiled
without this fix in the past.

* Back out erroneous DEVELOPMENT_TEAM values.

* Fix API nit

* Rely on MAIN_APP_BUNDLE_IDENTIFIER

* Rename .shared to .appGroup and make it a class var (modeling .standard)

* Move NetBasal calculations into its own class

* Move the app group suite name into a per-target Bundle extension

In addition, add the MAIN_APP_BUNDLE_IDENTIFIER into the Info.plist
so that each target can access it.

* Refactor the Bundle.appGroupSuiteName extension code for clarity

Eliminate duplicate code where we're choosing the appropriate bundle
inside the extension; instead let the caller choose the right bundle
and have common code for .appGroupSuiteName.

* Switch to a struct; use let instead of var

* Make all context struct members constant

This requires a slight restructuring of the way we initialize
GlucoseContext so that we create the SensorDisplayable first
if it exists.

* Set the Loop Status Extension build to CURRENT_PROJECT_VERSION