Only allow one active glucose data source
Separate CGM sources out into their own section and enforce zero or one
active sources at any given time. This requires a new boolean value to
track whether the G5 transmitter is enabled independently of the actual
transmitter ID.

While we're at it, create a separate config section for the pump to
group it with the pump battery type since those are not really
configuration values and go together.