Refactor Bolus recommendation code (#708)
* Refactor Bolus Recommendation

Make Bolus recommendation part of Loop update and don't allow
external calls to it.  The data doesn't change in any case and
update() is called in all places where we want a Bolus
recommendation.

This is in preparation of automated Bolus code, which needs
consistent Bolus and Basal data.

* Remove recommendBolus function from LoopState

The recommendedBolus variable contains the same information.