LOOP-1200: Persist Alerts to CoreData (#89)
* LOOP-1200: Persist DeviceAlerts to CoreData

Also: Squash removePendingAlert and removeDeliveredAlert into retractAlert

* PR Feedback: Renaming

* checkpoint

* Checkpoint: back to codegen'd CoreData object (seems to work)

* Remove bits about expiration until we know how/when we want to do it.

* Shuffled some code around a little.

* oops these changes weren't staged.

* Add unit tests, remove query support (for another PR)

* PR Feedback, more unit tests

* PR feedback

* Add a log line for when an alert is not found to update

* PR Feedback: "flatten" Trigger so that it is queryable

Stored trigger type as an Int16, trigger interval as an optional Double (NSNumber).

* Add query using modification counter (#92)

* Add query using modification counter

* PR Feedback

* LOOP-1200: Dump Alert database into Issue Report (#93)

* LOOP-1200: Dump Alert database into Issue Report

Added query by date, plus unit tests.

* PR Feedback, plus some rework/refactor

This takes @darin's suggestion, with some modifications :).  Now, the Filter is part of the Anchor, which makes it so anchoring continues to apply the filter as appropriate.

* LOOP-1200: alert persistence followups (#94)

* Add query using modification counter

* PR Feedback

* LOOP-1200: Dump Alert database into Issue Report

Added query by date, plus unit tests.

* PR Feedback, plus some rework/refactor

This takes @darin's suggestion, with some modifications :).  Now, the Filter is part of the Anchor, which makes it so anchoring continues to apply the filter as appropriate.

* LOOP-1200: Split alertIdentifier and managerIdentifier

After talking with Paul, we determined that we will need to be able to either filter or group alerts by device, so it would be better to keep managerIdentifier separate now so it can be used in data queries.