LOOP-940: Initial pass at new Settings Screen (#135)
* LOOP-940: Initial pass at new Settings Screen

So far, this is a pretty close approximation of the Settings screen as spec'd in LOOP-940.
It is not completely wired up; only the following are wired up:
- "Closed Loop" (dosingEnabled)
- alert permissions
- pump & cgm manager buttons

Note: This screen is currently hidden behind the "Alert Permissions" button in the "old" Settings UI.
Note also: because the old UI (SettingsTableViewController) still exists and has a lot of code logic in it, the way it
navigates to the device screens is by first dismissing and then presenting them.  This is temporary.

* wire up "Closed Loop" toggle

* Revert unneeded change

* minor change

* PR Feedback

* More PR Feedback.

* more PR feedback

* Remove unnecessary image assets

* Remove some unnecessary code

* Move Views out of LoopUI into Loop

* Changes based on design review feedback

* Black text on buttons

* Make "+" circle blue

* Fix the signaling from the composed viewmodels

Turns out, ObservableObjects don't automagically 'compose' such that when a composed OO changes, the composing OO changes.  This wires that up using `objectWillChange`.

* I didn't like how `LargeButton` "knew" how to render a "plus" button

...so I factored it out, and now it is only ever dealing with an "imageView".  (I wish I knew how to avoid the type erasure...)

* PR Feedback