LOOP-3853: Puts bolus error description into notification body (#455)
* LOOP-3853: Puts bolus error description into notification body if need be

Turns out the issue is that the `errorDescription` from the pump manager was being put into an iOS notification's `subtitle` field, which apparently can only be one line.  Thus, if it is too long, it gets ellipsized.
At first I just put it into the `body`, but then I thought that this might not be in the spirit of the original implementation, so I made a conservative estimate for the max length that would fit in the subtitle.  If it fits, it still puts it there.  If it doesn't, it puts it into the body.

* Reverts algorithm that was still trying to use subtitle.